// api/model/cardDto.ts
/**
 * CPMS API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { WorkspaceDto } from './workspaceDto';


export interface CardDto { 
    id?: string | null;
    tag?: string | null;
    expiryDate?: string | null;
    userId?: string | null;
    cardHolderName?: string | null;
    isBlocked?: boolean;
    directoryId?: string | null;
    workspaceId?: string | null;
    isActive?: boolean;
    directoryReadableId?: string | null;
    directoryName?: string | null;
    workspace?: WorkspaceDto;
}