// api/model/updateCardCommand.ts
/**
 * CPMS API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface UpdateCardCommand { 
    id?: string | null;
    workspaceId?: string | null;
    expiryDate?: string | null;
    cardHolderName?: string | null;
    directoryId?: string | null;
    isBlocked?: boolean | null;
    isArchived?: boolean | null;
    isActive?: boolean | null;
}